back/requirements.txt
django
django-ninja
django-cors-headers